car_classifier/data/features/car_features/requirements.txt
torch==1.7.1
torchvision==0.8.2
numpy
Pillow